Relocate pages to src/containers
